chore: add fake server for query-rest
The fake server returns search results and data. The search results do not really
search, so don't be surprised to find all sorts of nonsense in the responses.
Same goes for the generated data.

The responses are randomized in size and values, so the UI is a tad more
interesting and feels more interactive while developing.