build: get rid of triple-slash reference
For running tests through mocha and ts-node it was required to add a
triple-slash directive. That is something that eslint would object to.

Using the changes in the configurations ts-node will pick up on the
global.d.ts declarations and (compile and) run fine, even when started as a
module through mocha.